Continuing to work on #10
 - Continuing to add support for training;
 - Adding more CNTK backend implementations;
 - Updates #1: Contributing to Keras-Sharp;

+++ Sources/DataType.cs
@@ -11,7 +11,7 @@ namespace KerasSharp
 {
     public enum DataType : uint
     {
-        DEFAULT_DTYPE = Float,
+        DEFAULT_DTYPE = Double,
 
         Float = 1,
         Double = 2,
